Change project for activeadmin
