+++ app/assets/config/manifest.js
@@ -1,0 +1,2 @@
+//= link_directory ../../javascript .js
+//= link_directory ../../javascript .scss
