+++ app/controllers/api/v1/api_controller.rb
@@ -1,9 +1,20 @@
 module Api
   module V1
-    class ApiController < ApplicationController
+    class ApiController < ActionController::API
+      include DeviseTokenAuth::Concerns::SetUserByToken
+
+      before_action :configure_permitted_parameters, if: :devise_controller?
+      before_action :authenticate_user!, unless: :devise_controller?
+
       rescue_from ActiveRecord::RecordNotFound, with: :render_not_found
       rescue_from ActiveRecord::RecordInvalid, with: :render_record_invalid
       rescue_from ActionController::ParameterMissing, with: :render_parameter_missing
+
+      protected
+
+      def configure_permitted_parameters
+        devise_parameter_sanitizer.permit(:sign_up, keys: %w[gender])
+      end
 
       private
