+++ app/controllers/application_controller.rb
@@ -1,12 +1,3 @@
-class ApplicationController < ActionController::API
-  include DeviseTokenAuth::Concerns::SetUserByToken
-
-  before_action :configure_permitted_parameters, if: :devise_controller?
-  before_action :authenticate_user!, unless: :devise_controller?
-
-  protected
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: %w[gender])
-  end
+class ApplicationController < ActionController::Base
+  protect_from_forgery with: :exception
 end
